+++ jest.config.mjs
@@ -4,6 +4,7 @@ export default {
   ...createDefaultEsmPreset({ tsconfig: 'tsconfig.test.json' }),
   testEnvironment: 'node',
   testMatch: ['**/*.test.ts'],
+  testPathIgnorePatterns: ['<rootDir>/dist/'],
   moduleNameMapper: {
     '^(\\.{1,2}/.*)\\.js$': '$1',
     '^@harborclient/sdk/react$': 'react'