+++ src/client.d.ts
@@ -1,2 +1,2 @@
-export * from './index';
-export * from './runtime/index';
+export * from './index.js';
+export * from './runtime/index.js';